Refactor service and tests
